Release Beacon 2.0, beta 1.

+++ Beacon.framework/PrivateHeaders/HSBeaconKeyboardHeightObserver.h
@@ -2,7 +2,7 @@
 
 @protocol HSBeaconKeyboardHeightObserverDelegate
 
--(void)keyboardHeightChanged:(int)keyboardHeight;
+-(void)keyboardHeightChanged:(CGFloat)keyboardHeight;
 
 @end
